% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callmice-package.R
\docType{package}
\name{callmice}
\alias{callmice}
\alias{callmice-package}
\title{\pkg{callmice}: Test calling mice functions}
\description{
The \pkg{mice} package implements a method to deal with missing data.
The \pkg{callmice} package tests and demonstrates how to incorporate
functionality from \pkg{mice} into your own package.
}
\details{
The main function is \code{myimpute}.
}
\seealso{
\code{\link{myimpute}}
}
