% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myimpute.R
\name{myimpute}
\alias{myimpute}
\title{myimpute}
\usage{
myimpute(data = mice::nhanes, ...)
}
\arguments{
\item{data}{A \code{data.frame} with missing values.}

\item{\dots}{Other arguments passed to \code{mice}}
}
\value{
An object of class \code{mids}
}
\description{
This is an example function in your own package.
It calls `mice` and its underlying functions.
}
\examples{
imp <- myimpute()
}
\seealso{
\code{\link[mice]{mice}}
}
